refactor - Money 클래스내에 정적 팩토리 메서드를 생성해서 우승상금일때를 구분함
