+++ src/main/java/Lotto/domain/Money.java
@@ -4,9 +4,9 @@ import Lotto.exception.TooSmallMoneyException;
 
 public class Money {
 
-    private int money = 0;
     private static final int MIN_MONEY  = 1000;
 
+    private int money = 0;
 
     public Money(int money, boolean isWinningMoney) {
         if (!isWinningMoney) {
@@ -16,7 +16,11 @@ public class Money {
         this.money = money;
     }
 
-    private void validate(int money) {
+    public static Money createWinningMoney(int money) {
+        return new Money(money, true);
+    }
+
+    private static void validate(int money) {
         if (money < MIN_MONEY) {
             throw new TooSmallMoneyException("구매금액은 최소" + MIN_MONEY + "이상이어야합니다.");
         }
